Move ConnectorIcon and DefaultLogoCatalog to components/common
