--- airbyte-webapp/src/components/DefaultLogoCatalog/index.tsx
@@ -1,4 +1,0 @@
-import DefaultLogoCatalog from "./DefaultLogoCatalog";
-
-export * from "./DefaultLogoCatalog";
-export { DefaultLogoCatalog };
