+++ airbyte-webapp/src/components/EntityTable/components/ConnectorCell.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { ConnectorIcon } from "components/ConnectorIcon";
+import { ConnectorIcon } from "components/common/ConnectorIcon";
 
 interface IProps {
   value: string;
