+++ airbyte-webapp/src/components/EntityTable/components/NameCell.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from "react";
 import { useIntl } from "react-intl";
 import styled from "styled-components";
 
-import { ConnectorIcon } from "components/ConnectorIcon";
+import { ConnectorIcon } from "components/common/ConnectorIcon";
 import StatusIcon from "components/StatusIcon";
 import { StatusIconStatus } from "components/StatusIcon/StatusIcon";
