+++ airbyte-webapp/src/components/common/ConnectorIcon/ConnectorIcon.tsx
