+++ airbyte-webapp/src/components/common/ConnectorIcon/index.ts
