+++ airbyte-webapp/src/components/common/DefaultLogoCatalog/DefaultLogoCatalog.tsx
@@ -1,4 +1,4 @@
-const DefaultLogoCatalog = (): JSX.Element => (
+export const DefaultLogoCatalog = (): JSX.Element => (
   <svg width="18" height="18" viewBox="0 0 61 61" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path
       d="M59.8281 46.2031L45.2969 31.5547C43.4219 29.7969 41.0781 28.8594 38.6172 28.8594C36.8594 28.8594 34.9844 29.3281 33.4609 30.3828L23.5 20.4219V11.75L8.5 0.5L1 8L12.25 23H20.8047L30.7656 33.0781C28.4219 36.7109 28.7734 41.6328 32.0547 44.9141L46.5859 59.4453C47.2891 60.1484 48.3438 60.6172 49.2812 60.6172C50.2188 60.6172 51.1562 60.1484 51.8594 59.4453L59.8281 51.4766C61.3516 50.0703 61.3516 47.6094 59.8281 46.2031ZM19.75 19.1328V19.25H14.125L5.92188 8.35156L8.85156 5.53906L19.75 13.625V19.1328ZM49.2812 56.75L34.6328 42.2188C33.5781 41.1641 32.9922 39.7578 32.9922 38.2344C32.9922 36.7109 33.5781 35.3047 34.6328 34.25C35.6875 33.1953 37.0938 32.6094 38.6172 32.6094C40.1406 32.6094 41.5469 33.1953 42.6016 34.25L57.25 48.8984L49.2812 56.75ZM8.5 51.125C8.5 52.1797 9.32031 53 10.375 53C11.3125 53 12.25 52.1797 12.25 51.125C12.25 50.1875 11.3125 49.25 10.375 49.25C9.32031 49.25 8.5 50.1875 8.5 51.125ZM33.3438 8.35156C36.0391 5.77344 39.5547 4.25 43.1875 4.25C44.0078 4.25 44.8281 4.36719 45.6484 4.48438L37.5625 12.6875L39.0859 22.4141L48.8125 23.9375L57.0156 15.8516C57.7188 20.3047 56.3125 24.875 53.1484 28.1562C52.0938 29.2109 50.9219 30.0312 49.6328 30.6172L52.3281 33.4297C53.6172 32.7266 54.6719 31.7891 55.7266 30.7344C60.1797 26.2812 61.9375 20.0703 60.4141 13.9766C60.1797 12.8047 59.2422 11.9844 58.1875 11.6328C57.0156 11.3984 55.8438 11.6328 55.0234 12.4531L47.5234 19.9531L42.3672 19.1328L41.5469 13.9766L49.0469 6.47656C49.75 5.65625 50.1016 4.48438 49.8672 3.3125C49.5156 2.25781 48.5781 1.32031 47.5234 1.08594C41.4297 -0.4375 35.2188 1.32031 30.7656 5.77344C29.2422 7.29688 28.0703 9.05469 27.25 10.9297V18.8984L29.8281 21.4766C28.6562 16.6719 29.9453 11.75 33.3438 8.35156ZM13.5391 55.3438C11.5469 57.2188 8.14844 57.2188 6.15625 55.3438C4.16406 53.2344 4.16406 49.9531 6.15625 47.9609L23.2656 30.8516L20.6875 28.2734L3.57812 45.3828C0.0625 48.7812 0.0625 54.5234 3.57812 57.9219C5.21875 59.6797 7.44531 60.5 9.78906 60.5C12.25 60.5 14.4766 59.6797 16.1172 57.9219L28.0703 45.9688C27.25 44.9141 26.6641 43.7422 26.1953 42.5703L13.5391 55.3438Z"
@@ -6,5 +6,3 @@ const DefaultLogoCatalog = (): JSX.Element => (
     />
   </svg>
 );
-
-export default DefaultLogoCatalog;
