+++ airbyte-webapp/src/components/common/DefaultLogoCatalog/index.ts
@@ -1,0 +1,1 @@
+export * from "./DefaultLogoCatalog";
