+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/components/ConnectionCell.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
 import styled from "styled-components";
 
-import { ConnectorIcon } from "components/ConnectorIcon";
+import { ConnectorIcon } from "components/common/ConnectorIcon";
 
 interface ConnectionCellProps {
   sourceDefinitionName: string;
